More command work.

+++ packages/cli/src/commands/Init.ts
@@ -1,7 +1,7 @@
 import fs from 'fs-extra';
 import { Path } from '@boost/common';
 import { Arg, Config, Command, GlobalOptions } from '@boost/cli';
-import { CONFIG_FILE } from '../constants';
+import { CONFIG_FILE, CWD } from '../constants';
 
 export interface InitOptions extends GlobalOptions {
   modularScale: boolean;
@@ -17,18 +17,25 @@ export default class Init extends Command<InitOptions, [string]> {
     label: 'cwd',
     type: 'string',
   })
-  async run(cwd: string = process.cwd()) {
-    const configPath = Path.resolve(CONFIG_FILE, cwd);
+  async run(cwd: string = CWD) {
+    const targetPath = new Path(cwd);
+    const configPath = targetPath.append(CONFIG_FILE);
     const templatePath = new Path(
       require.resolve(
         `@aesthetic/compiler/templates/${this.modularScale ? 'config' : 'config-fixed'}.yaml`,
       ),
     );
 
-    this.log('Creating config file in %s', cwd);
+    if (!targetPath.exists()) {
+      throw new Error(`Target directory "${targetPath}" does not exist.`);
+    }
+
+    this.log('Creating config file in %s', targetPath);
 
     await fs.copyFile(templatePath.path(), configPath.path());
 
-    this.log('Created!');
+    this.log('Created %s!', CONFIG_FILE);
+
+    // TODO Add enquirer integration and ask basic questions to inject into the yaml file
   }
 }
